LayoutTests/http/tests/workers/service/service-worker-cache-api.https-expected.txt

Registering service worker
Service worker registered
Status is cache is ready
Fetching
Response buffer byte length is 12940
Status is matched cache for https://127.0.0.1:8443/resources/square100.png
